Update README.md
Passing Arguments to Functions

Rename the function myFunction so that it has a sensible name. Don't forget to rename it where it is called.
Call the function with the arguments 'Daffy' and 'Duck'.
Ask the user to seperately input a first and last name. Store their inputs in two variables. Now call the function with the variables as the arguments.